Fix literal string escaping
svgbob will render strings in double quotes literally, and not replace any
characters by svg paths.

However, as mentioned in #19, there is currently a bug that the last part of a
line after the last closing double quote character is ignored:

    +-----------+---+
    | "a.to(b)" | c |
    +-----------+---+

renders like

    +-----------+---+
    | "a.to(b)"
    +-----------+---+

which is fixed in this pull request.