Make --scale a multiplier of the default scale
The help for the `--scale` option states the default value is 1. This is not the case though (current default is 8.0) which leads to the surprising result that without a scale argument the output width will be `x` and with, for instance `--scale 2` the output width will be `x / 4` instead of `x * 2`.

This change makes the CLI scale argument a multiplier of the default scale instead of overwriting the default scale.